requirements.txt
asgiref==3.6.0 ; python_version >= '3.7'
certifi==2022.12.7 ; python_version >= '3.6'
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.0.1
crispy-bootstrap5==0.7
cryptography==39.0.1
defusedxml==0.7.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
django==4.1.7
django-allauth==0.52.0
django-crispy-forms==2.0 ; python_version >= '3.7'
django-environ==0.9.0
django-resized==1.0.2
idna==3.4 ; python_version >= '3.5'
oauthlib==3.2.2 ; python_version >= '3.6'
pillow==9.4.0
psycopg2-binary==2.9.5
pycparser==2.21
pyjwt[crypto]==2.6.0 ; python_version >= '3.7'
python3-openid==3.2.0
requests==2.28.2 ; python_version >= '3.7' and python_version < '4'
requests-oauthlib==1.3.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.3 ; python_version >= '3.5'
tzdata==2022.7 ; sys_platform == 'win32'
urllib3==1.26.14 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'